+++ packages/modal/src/components/modal/ModalHeader.tsx
@@ -4,8 +4,8 @@ import cx from "classnames";
 import * as React from "react";
 import { ReactNode } from "react";
 import { DRAGGABLE_HANDLE_CLASSNAME } from "./BaseModal";
-
 import styles from "./Modal.module.css";
+import { Box } from "@stenajs-webui/core";
 
 interface Props {
   onRequestClose?: () => void;
@@ -22,7 +22,9 @@ export const ModalHeader: React.FC<Props> = ({
     <div className={cx(styles.header, DRAGGABLE_HANDLE_CLASSNAME)}>
       {headerText}
       {header}
-      <FlatButton onClick={onRequestClose} leftIcon={faTimes} />
+      <Box alignSelf={"flex-start"}>
+        <FlatButton onClick={onRequestClose} leftIcon={faTimes} />
+      </Box>
     </div>
   );
 };