+++ CHANGELOG.md
@@ -1,9 +1,9 @@
 # Changelog
 
-## Upcoming
+## Upcoming changes
 
-### design changes
-
+### Design changes
+- `Banner` icon has fixed width and will not be compressed by long content
 - All horizontal column alignments will be rendered according to the table config
 
 ## 9.0.0