+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ### Design changes
 - `Banner` icon has fixed width and will not be compressed by long content
+- Button variants look good for `SecondaryButton` and `FlatButton`
+- Border color matches background color for `PrimaryButton`
+- Disabled text and icon color for `PrimaryButton` is the same as for `SecondaryButton`
 
 ## 9.0.0