+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Upcoming changes
+
+### Modal
+
+- Close button is aligned to top of modal header
+
 ## 9.1.0
 
 ### Select