+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Upcoming changes
 
+### Modal
+
+- Close button is aligned to top of modal header
+
 ### ValueTable
 
 - Adds `variant` prop with values "relaxed", "standard", "condensed", "compact".